fix issues

+++ utils/expense_calculator.py
@@ -1,17 +1,17 @@
 class Calculator:
     @staticmethod
-    def multiply(a: int, b: int) -> int:
+    def multiply(a: float, b: float) -> float:
         """
         Multiply two numbers.
         
         Args:
-            a (int): The first number.
-            b (int): The second number.
+            a (float): The first number.
+            b (float): The second number.
         
         Returns:
-            int: The product of a and b.
+            float: The product of a and b.
         """
-        return a * b
+        return float(a) * float(b)
     
     @staticmethod
     def calculate_total(*x: float) -> float:
